#187: consolidate formatBytes to commonMain, drop material3 dep, doc/config cleanups
