+++ composeApp/build.gradle.kts
@@ -96,7 +96,6 @@ kotlin {
             implementation(libs.kydra.log)
             implementation(libs.compose.runtime)
             implementation(libs.compose.foundation)
-            implementation(libs.compose.material3)
             implementation(libs.compose.ui)
             implementation(libs.compose.components.resources)
             implementation(libs.compose.uiToolingPreview)
