+++ composeApp/src/desktopCli/kotlin/com/tubetoast/tether/Main.kt
@@ -14,6 +14,7 @@ import com.tubetoast.tether.network.FileClient
 import com.tubetoast.tether.network.send
 import com.tubetoast.tether.protocol.Device
 import com.tubetoast.tether.protocol.SendResult
+import com.tubetoast.tether.util.formatBytes
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -200,13 +201,6 @@ suspend fun handleSend(
         is SendResult.Success -> output("[send] OK — ${elapsed.inWholeMilliseconds} ms  →  ${result.savedPath}")
         is SendResult.Failure -> output("[send] FAIL: ${result.reason}")
     }
-}
-
-fun formatBytes(bytes: Long): String = when {
-    bytes < 1_024 -> "$bytes B"
-    bytes < 1_024 * 1_024 -> "%.1f KB".format(bytes / 1_024.0)
-    bytes < 1_024 * 1_024 * 1_024 -> "%.1f MB".format(bytes / (1_024.0 * 1_024))
-    else -> "%.2f GB".format(bytes / (1_024.0 * 1_024 * 1_024))
 }
 
 fun parseTokens(line: String): List<String> {
